Created function for fetching the value of a
single cell from the board.

Maybe I will rebuild the other accessors to be built on top of this.
At the least, I need to bring some more coherence into the API as to
return values; right now it is a mix of Maybe values and unsafe values.
I am thinking that safety is worth dealing with Maybe. And perhaps I
could make unsafe versions of each for times when safety is predetermined.